backend/requirement.txt
# =========================
# Core Framework
# =========================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0

# =========================
# Environment & Configuration
# =========================
python-dotenv==1.0.0

# =========================
# File & Async Operations
# =========================
aiofiles==23.2.1

# =========================
# HTTP / API
# =========================
requests==2.31.0
groq==0.9.0

# =========================
# Audio Processing (Whisper)
# =========================
openai-whisper==20231117
soundfile==0.12.1
numpy==1.24.3

# =========================
# NLP / Text
# =========================
markdown==3.5.2

# =========================
# Document Generation
# =========================
fpdf==1.7.2
